feat: add save component
